PDP-1139: renamed features
