+++ lib/average_end_of_day_depository_balance/.env.local
